feat(console): Refactor code with class 'Console'
